requirements.txt
pytest
paramiko
pyyaml
